Move serialization implementation and rename messagepack feature to serialize_rmp
