+++ src/events/payload_serializer/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "serialize_rmp")]
+pub mod serialize_rmp;
+#[cfg(feature = "serialize_rmp")]
+pub use serialize_rmp::*;
